perf(printer): dedupe image scan summary in O(N) via a shared helper
